rename usize/isize to uint/int; prefix C typedefs with __sl_
- Rename SL builtin types: usize → uint, isize → int
- Change C codegen to emit __sl_-prefixed typedefs in EmitPrelude
  (e.g. typedef uint8_t __sl_u8) so generated headers don't conflict
  with host code using names like u8, i32, bool, etc., and to allow
  typedef ptrdiff_t __sl_int (int is a C keyword)
- Split builtinNames[] into builtinSlNames[]/builtinCNames[] in
  ResolveTypeName so SL names map to their __sl_ C counterparts
- Update all BufAppendCStr emission sites to use __sl_ type names
- Update docs, grammar examples, tests, and build.sh accordingly

+++ src/slc.c
@@ -1162,8 +1162,8 @@ static int IsBuiltinTypeName(const char* src, uint32_t start, uint32_t end) {
         || SliceEqCStr(src, start, end, "u16") || SliceEqCStr(src, start, end, "u32")
         || SliceEqCStr(src, start, end, "u64") || SliceEqCStr(src, start, end, "i8")
         || SliceEqCStr(src, start, end, "i16") || SliceEqCStr(src, start, end, "i32")
-        || SliceEqCStr(src, start, end, "i64") || SliceEqCStr(src, start, end, "usize")
-        || SliceEqCStr(src, start, end, "isize") || SliceEqCStr(src, start, end, "f32")
+        || SliceEqCStr(src, start, end, "i64") || SliceEqCStr(src, start, end, "uint")
+        || SliceEqCStr(src, start, end, "int") || SliceEqCStr(src, start, end, "f32")
         || SliceEqCStr(src, start, end, "f64");
 }
